new star component

+++ app/src/main/java/com/uri/bolanope/model/RatingModel.kt
@@ -1,5 +1,5 @@
 package com.uri.bolanope.model;
 
 data class RatingModel(
-    val average_rating: Int
+    val average_rating: Float
 )
